feat: Include data warehouse tables on web overview
WebOverview now includes Data Warehouse tables currency data. It's a bit sketchy because they don't include `session_id` so we need to be slightly smart on how we build it.

Also, for tests, this sucks. We want the snapshots to be properly generated, BUT, it's very hard to run against a data warehouse table because it doesn't exist on S3. I might make it work with CSV soon.